Changed the update method to be consistent between all bases.

+++ Cocos2DGame/Base.h
@@ -17,7 +17,7 @@
 @property (nonatomic, assign) BOOL selected;
 @property (nonatomic, assign) ccColor3B defaultColor;
 @property (nonatomic, assign) id<BaseDelegate> delegate;
-@property (nonatomic, assign) int count;
+@property (nonatomic, assign) float count;
 @property (nonatomic, assign) int capacity;
 @property (nonatomic, retain) CCLabelTTF *label;
 @property (nonatomic, assign) Team team;
